AI Update (levels 1 and 3 working)
In addition to making levels one and three work for the AI,
I also fixed a bug where the game would kick you from the round
if you only had 3s and/or 4s. There was also a bug where computer
players could reach values higher than 99 when set to level 1.
Aesthetically, I made a note when a player's turn is skipped.